refactor: prefix stake neuron component with Nns
