+++ frontend/src/lib/components/neurons/NnsStakeNeuron.svelte
